+++ lib/helpers/dependencies_injection.dart
@@ -4,8 +4,11 @@
 import 'package:get_it/get_it.dart';
 import 'package:xiaokeai/helpers/get_platform_service.dart';
 import 'package:xiaokeai/services/auth/acc/auth_service.dart';
+import 'package:xiaokeai/services/auth/acc/auth_wrapper.dart';
 import 'package:xiaokeai/services/auth/local/local_auth_service.dart';
 import 'package:xiaokeai/services/configs/utils/permission_service.dart';
+import 'package:xiaokeai/services/object_storage/cloud_object_storage_service.dart';
+import 'package:xiaokeai/services/object_storage/cloud_object_storage_wrapper.dart';
 import 'package:xiaokeai/services/package/package_info_provider.dart';
 import 'package:xiaokeai/services/package/package_info_service.dart';
 import 'package:xiaokeai/services/pref/shared_pref_service.dart';
@@ -14,8 +17,7 @@ final GetIt getIt = GetIt.instance;
 
 Future<void> setupDependencies() async {
   setupAuthServices();
-  // getIt.registerLazySingleton(() => CloudObjectStorageService());
-  // getIt.registerLazySingleton(() => CloudObjectStorageWrapper());
+  setupCloudObjectStorageServices();
   // getIt.registerLazySingleton(() => FirestoreService());
   // getIt.registerLazySingleton(() => FirestoreWrapper());
   getIt.registerLazySingleton(() => PackageInfoService());
@@ -28,6 +30,11 @@ Future<void> setupDependencies() async {
   // getIt.registerLazySingleton(() => SpeechService());
 }
 
+void setupCloudObjectStorageServices() {
+  getIt.registerLazySingleton(() => CloudObjectStorageService());
+  getIt.registerLazySingleton(() => CloudObjectStorageWrapper());
+}
+
 Future<void> setupSharedPref() async {
   final sharedPrefsService = SharedPreferencesService();
   await sharedPrefsService.init();
@@ -36,5 +43,5 @@ Future<void> setupSharedPref() async {
 
 void setupAuthServices() {
   getIt.registerLazySingleton(() => AuthService());
-  // getIt.registerLazySingleton(() => AuthWrapper());
+  getIt.registerLazySingleton(() => AuthWrapper());
 }