+++ lib/shared/common/file_picker_enum.dart
@@ -1,0 +1,15 @@
+/// lib/shared/common/file_picker_enum.dart
+///
+/// enum for picker mode
+
+enum FilePickerMode {
+  /// meant to support only image file type
+  ///
+  /// opens the platform photo library
+  photoPicker,
+
+  /// meant to support any file type
+  ///
+  /// opens the platform file picker
+  filePicker,
+}